Hide stderror during gl swap call
