+++ apps/vaporgui/VizWin.cpp
@@ -50,6 +50,7 @@
 #include <vapor/FlowParams.h>
 #define INCLUDE_DEPRECATED_LEGACY_VECTOR_MATH
 #include <vapor/LegacyVectorMath.h>
+#include "hide_std_error_util.h"
 
 using namespace VAPoR;
 
@@ -683,7 +684,9 @@ void VizWin::Render(bool fast) {
 #endif
     }
 	
+    HideSTDERR();
 	swapBuffers();
+    RestoreSTDERR();
     
 	rc = CheckGLErrorMsg("VizWindowPaintGL");
 	if (rc < 0) {
