+++ apps/vaporgui/hide_std_error_util.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void HideSTDERR();
+void RestoreSTDERR();
